Improve tool icon stylings and widget intilization
The toolbar is not styled to snap to the right side of the screen and
resize smaller as the browser resolution gets smaller.  When a tool is
open, the associated dialogue box now opens in a defined location on the
left side of the screen.  The latter update will improve UX on mobile
devices.